+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @elf-framework/design-tokens
 
+## 0.0.53
+
+### Patch Changes
+
+- update design token values
+
+## 0.0.50
+
+### Patch Changes
+
+- update dark color in button
+
 ## 0.0.49
 
 ### Patch Changes