+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @elf-framework/ui
 
+## 0.0.53
+
+### Patch Changes
+
+- update design token values
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.53
+
+## 0.0.52
+
+### Patch Changes
+
+- fixed inputPaint value display
+
+## 0.0.50
+
+### Patch Changes
+
+- update dark color in button
+
 ## 0.0.49
 
 ### Patch Changes